requirements.txt
flask
flask-cors
pandas
requests
gspread
oauth2client
gspread_dataframe
dateparser
undetected-chromedriver
selenium
python-dateutil
yfinance
google-auth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
webdriver-manager
setuptools
